Shortened Vim's native keypress pause/wait time.
Fun fact: Vim also supports mappings that include key codes (eg.
<Enter>, <Del>, etc...). Thus, there's a native ~1s wait following a key
code to support the fact that you *may* be typing a mapping, for
instance.

Particularly, this is super annoying in the case that you've just exited
INSERT mode by pressing <ESC>, followed by an immediate Shift+O to
re-enter INSERT mode on the line above, as there's a very noticeable
lag.

Rather than turn this off for <ESC> specifically with 'noesckeys', this
commit effectively segments out the timeout for general mappings and the
timeout (ttimeout) for key codes specifically, the latter of which is
drastically reduced (100ms). This gives the benefit of still allowing a
reasonable wait time for mappings specifically, while "fixing" the key
code scenario outlined above.