Enabled matchit.vim.
For a while now, the plugin 'matchit.vim' has shipped with core Vim.
However, it's not enabled by default. This commit does so.

matchit.vim greatly improves/enhances the '%' character to support
moving between, for instance, HTML tags (as opposed to the default which
only moves between '<' and '>' within the same tag) in addition to a
slew of other things (eg. more intelligent behavior with comments).