Added a mapping to clear hlsearch w/o jibberish.
It's definitely preferable to have 'hlsearch' enabled, but having to
type a random string to clear it when you're done is obviously annoying
and/or suboptimal. This commit adds a mapping in normal mode that clears it in a very
intuitive way (by hitting <ESC>).

For doing so, we use ':nohlsearch' -- NOTE: this is explicitly different
from ':set nohlsearch'. In the former (what we're using), search
highlighting is only disabled temporarily and is automatically
re-enabled on subsequent searches; in the latter, search highlighting is
turned off until Vim restarts.

FWIW, I worry a little bit about this mapping conflicting with other
<ESC>-related stuff, but until there's an actual manifestation of a
problem, it's working great as-is.