Recreated solid .osx defaults script from scratch.
Although not entirely "complete" (still WIP), this commit represents a
solid enough foundation on which to build going forward.

I previously had an `~/.osx` defaults script entirely based on
`https://github.com/mathiasbynens/dotfiles/blob/master/.osx`, but having
not spent a ton of time making modifications, there were way too many
settings that I wasn't particularly fond of and caused undesirable
side-effects. Further, once run, there wasn't a great way of reverting
to defaults and cherry-picking just the "good" mutations.

This time around, I went through with a fine-toothed comb and extracted
exactly the settings/defaults that I agreed with, along with many of my
own pulled from my local `defaults` domain(s), and in combination with a
host of others pulled from popular dotfiles repositories across the web.

Eventually, I'd like to get to the point that, upon a fresh OSX install,
zero explicit configuration is required for any/all of the apps that I
use on a daily basis (as it'd be embedded into this script), but there's
a lot left to be desired to-date in the way of that as an end-goal.