Flushed installed binaries into a fresh Brewfile.
Used Homebrew's `bundle brew` to, having just reinstalled OSX and
reinitialized my dev environment with all of the necessary binaries and
executables, flush what I'd installed into a Brewfile such that it can
be automated the next time around.

What's included is everything that I consider essential, along with some
that are commented out because they could be considered closer to
optional, depending on the work that's being performed.

Preferably, I'd like to err on the side of containerizing
anything/everything that can be (in the interest of tidiness),
especially more infrastructure-related project dependencies (eg. `lein`,
`node`, `go`, etc...).