Added vim-repeat for more robust `.` repetition.
Specifically, with something like vim-sexp where you're moving elements
and forms around, it helps to only have to type it once and be able to
repeatedly press the `.` key.

However, *many* plugins have support for vim-repeat (or vice-versa,
whichever), so it's a quick win in that you get a lot of added
repeat-based behavior across the board for free!