Started public README with basic instructions.
I have a much more in-depth wiki that I'm developing locally for myself,
but it includes far too much specific and/or private information related
to my particular setup / use-case to be useful.

The purpose of this document is just so that people know what they're
looking at and/or how it can be used or tailored to their own use case!