Added augroup, more intuitive cursorline behavior.
I generally like `cursorline` because it makes it very easy to find the
cursor and focus on the given line, but the problem I was having was,
with several splits open (especially horizontal), there were cursorlines
everywhere, and it made it very visually distracting and hard to tell
which pane had focus. What's more, at least using my current
colorscheme, `cursorline` looks a lot like the statusline, so that made
it even more visually confusing at a glance.

This augroup effectively only enables `cursorline` for the focused pane.