Replaced custom nvim config with stock AstroNvim.
After many years, I've come to the realization that I don't have the
time or patience to manage my own (N)Vim configuration from scratch,
especially given that Nvim and the plugin ecosystem is a moving target
(for better *and* worse!).

Instead, it makes more sense to defer to the wisdom of the community for
the base configuration boilerplate, whereas I'll make selective,
explicit overrides if/when necessary. That way, I'm not constantly
having to set aside a weekend every 6 months when a new version of
Neovim is released, a better plugin comes along and/or an existing one
stops being maintained, etc...

To that end, I tried AstroNvim locally among several other alternatives,
and having used it with great success for many months, I'm willing to
now commit!