Remapped laptop keyboard only, Super<-->Alt.
Because my external mechanical keyboard has hardware-level remapping in
place already, let's just override the laptop's own keyboard to swap
Super (Mod) and Alt so that both behave predictably based on muscle
memory.

EDIT, addendum: per some testing, the above more general directive
doesn't fall through to the device-specific directive, so I ended up
having to be a bit redundant. There may be a way to refactor this in the
future? For now, it does the job.